requirements.txt

asn1crypto==0.24.0
astunparse==1.6.3
beautifulsoup4==4.6.0


cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
click==6.7
colorama==0.3.7
cryptography==2.1.4

cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==4.1.2



filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gdown==3.13.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0

grpcio==1.32.0
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==3.2
importlib-metadata==4.6.4

joblib==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0

launchpadlib==1.10.6
lazr.restfulclient==0.13.5


lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4

matplotlib==2.1.1
numpy==1.19.5



opt-einsum==3.3.0
pandas==0.22.0
pbr==5.6.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73

PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0


python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0



tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-addons==0.13.0
tensorflow-estimator==2.4.0
tensorrt
termcolor==1.1.0
testresources==2.0.1
threadpoolctl==3.1.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.0
typeguard==2.12.1
typing-extensions==3.7.4.3




urllib3==1.26.6
urwid==2.0.1
wadllib==1.3.2
webencodings==0.5
Werkzeug==2.0.1


youtube-dl==2018.3.14
zipp==3.5.0
zope.interface==4.3.2
